+++ packages/browser-wallet/src/popup/pages/VerifiableCredential/VerifiableCredentialList.tsx
@@ -7,7 +7,12 @@ import {
 import { useAtomValue, useAtom } from 'jotai';
 import Topbar, { ButtonTypes } from '@popup/shared/Topbar/Topbar';
 import { useTranslation } from 'react-i18next';
-import { VerifiableCredential, VerifiableCredentialSchema, VerifiableCredentialStatus } from '@shared/storage/types';
+import {
+    VerifiableCredential,
+    VerifiableCredentialSchema,
+    VerifiableCredentialSchemaWithFallback,
+    VerifiableCredentialStatus,
+} from '@shared/storage/types';
 import {
     VerifiableCredentialMetadata,
     getChangesToCredentialMetadata,
@@ -77,7 +82,7 @@ export function VerifiableCredentialCardWithStatusFromChain({
     className: string;
     onClick?: (
         status: VerifiableCredentialStatus,
-        schema: VerifiableCredentialSchema,
+        schema: VerifiableCredentialSchemaWithFallback,
         metadata: VerifiableCredentialMetadata,
         localization?: Record<string, string>
     ) => void;
@@ -120,7 +125,7 @@ export default function VerifiableCredentialList() {
     const [selected, setSelected] = useState<{
         credential: VerifiableCredential;
         status: VerifiableCredentialStatus;
-        schema: VerifiableCredentialSchema;
+        schema: VerifiableCredentialSchemaWithFallback;
         metadata: VerifiableCredentialMetadata;
         localization?: Record<string, string>;
     }>();
@@ -192,7 +197,7 @@ export default function VerifiableCredentialList() {
                             credential={credential}
                             onClick={(
                                 status: VerifiableCredentialStatus,
-                                schema: VerifiableCredentialSchema,
+                                schema: VerifiableCredentialSchemaWithFallback,
                                 metadata: VerifiableCredentialMetadata,
                                 localization?: Record<string, string>
                             ) => setSelected({ credential, status, schema, metadata, localization })}