+++ packages/browser-wallet/src/shared/storage/types.ts
@@ -345,3 +345,5 @@ export interface VerifiableCredentialSchema {
     properties: SchemaProperties;
     required: string[];
 }
+
+export type VerifiableCredentialSchemaWithFallback = VerifiableCredentialSchema & { usingFallback: boolean };